+++ protocols/kad/src/kbucket/bucket.rs
@@ -504,18 +504,15 @@ mod tests {
                     value: (),
                 };
                 let full = bucket.num_entries() == K_VALUE.get();
-                match bucket.insert(node, status) {
-                    InsertResult::Inserted => {
-                        let vec = match status {
-                            NodeStatus::Connected => &mut connected,
-                            NodeStatus::Disconnected => &mut disconnected,
-                        };
-                        if full {
-                            vec.pop_front();
-                        }
-                        vec.push_back((status, key.clone()));
+                if let InsertResult::Inserted = bucket.insert(node, status) {
+                    let vec = match status {
+                        NodeStatus::Connected => &mut connected,
+                        NodeStatus::Disconnected => &mut disconnected,
+                    };
+                    if full {
+                        vec.pop_front();
                     }
-                    _ => {}
+                    vec.push_back((status, key.clone()));
                 }
             }
 
@@ -533,7 +530,7 @@ mod tests {
             // All nodes before the first connected node must be disconnected and
             // in insertion order. Similarly, all remaining nodes must be connected
             // and in insertion order.
-            nodes == Vec::from(disconnected) && tail == Vec::from(connected)
+            disconnected == nodes && connected == tail
         }
 
         quickcheck(prop as fn(_) -> _);
@@ -635,7 +632,7 @@ mod tests {
 
         // The pending node has been discarded.
         assert!(bucket.pending().is_none());
-        assert!(bucket.iter().all(|(n, _)| &n.key != &key));
+        assert!(bucket.iter().all(|(n, _)| n.key != key));
 
         // The initially disconnected node is now the most-recently connected.
         assert_eq!(